Created folder for packet, added fragment packet (inprogress)
