+++ src/packet/type.rs
@@ -1,0 +1,5 @@
+pub enum PacketType
+{
+    PACKET_FRAGMENT = 0,
+
+}
